requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
anyio==4.0.0
arxiv==1.4.8
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.99.1
feedparser==6.0.10
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.34
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.57.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain==0.0.279
langsmith==0.0.33
loguru==0.7.0
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai @ git+https://github.com/openai/openai-python.git@1be14ee34a0f8e42d3f9aa5451aa4cb161f1781f
orjson==3.9.5
outcome==1.2.0
packaging==23.1
pandas==2.1.0
Pillow==9.5.0
pinecone-client==2.2.2
protobuf==3.20.3
pyarrow==13.0.0
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.2
rpds-py==0.10.2
selenium==4.12.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.20
starlette==0.27.0
streamlit==1.26.0
tenacity==8.2.3
tiktoken==0.3.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
watchfiles==0.20.0
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
